fix: moved all latex rendering into latex_artifact_generator.py for better origanization
