+++ app/latex/latex_artifact_generator.py
