+++ app/services/exam_rendering/latex_exam_pdf_rendering.py
@@ -5,7 +5,7 @@ import re
 from pathlib import Path
 from typing import Any
 
-from app.services.latex_image_service import build_latex_document_pdf_at_path
+from app.latex.latex_artifact_generator import build_latex_document_pdf_at_path
 
 logger = logging.getLogger(__name__)
