+++ app/services/messaging_service.py
@@ -7,6 +7,11 @@ import app.database.db as db
 import app.database.enums as enums
 import app.database.models as models
 from app.config import llm_settings
+from app.latex.latex_artifact_generator import (
+    looks_like_latex,
+    prepare_latex_body,
+    text_to_img,
+)
 from app.monitoring.metrics import record_messages_generated, track_messages
 from app.services.agent_client import agent_client
 from app.services.exam_delivery_service import (
@@ -15,11 +20,6 @@ from app.services.exam_delivery_service import (
     exam_delivery_service,
 )
 from app.services.flows.flow_service import flow_client
-from app.services.latex_image_service import (
-    looks_like_latex,
-    prepare_latex_body,
-    text_to_img,
-)
 from app.services.llm_service import llm_client
 from app.services.whatsapp_service import DocumentType, ImageType, whatsapp_client
 from app.tools.registry import ToolName
